*: move pkg/flags to pkg/localdata
